Clean up metrics handler imports
